Add generatePokeBg function from generatePokemon.js to generatePokeBg.js file
