+++ src/modules/generatePokeBg.js
@@ -63,4 +63,21 @@ const generateColor = async (type) => {
   return color;
 };
 
-export default generateColor;
+const generatePokeBg = async (types) => {
+  let startColor = '';
+  let endColor = '';
+  if (types.length === 2) {
+    startColor = await generateColor(types[0].type.name);
+    endColor = await generateColor(types[1].type.name);
+  } else if (types.length === 1) {
+    startColor = await generateColor(types[0].type.name);
+    endColor = startColor;
+  } else {
+    startColor = '#fff';
+    endColor = '#fff';
+  }
+  const result = await `background: linear-gradient(${startColor}, ${endColor})`;
+  return result;
+};
+
+export default generatePokeBg;
