+++ src/modules/generatePokemon.js
@@ -1,24 +1,7 @@
-import generateColor from './generatePokeBg.js';
+import generatePokeBg from './generatePokeBg.js';
 import getPokemonData from './getPokemonData.js';
 
 const pokiPopup = document.querySelector('#pokiPopup');
-
-const generatePokeBg = async (types) => {
-  let startColor = '';
-  let endColor = '';
-  if (types.length === 2) {
-    startColor = await generateColor(types[0].type.name);
-    endColor = await generateColor(types[1].type.name);
-  } else if (types.length === 1) {
-    startColor = await generateColor(types[0].type.name);
-    endColor = startColor;
-  } else {
-    startColor = '#fff';
-    endColor = '#fff';
-  }
-  const result = await `background: linear-gradient(${startColor}, ${endColor})`;
-  return result;
-};
 
 const generateTypesHtml = async (types) => {
   let generateHtml = '';
